% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findS3Meth.R
\name{findS3Meth}
\alias{findS3Meth}
\title{Find S3 Methods That Apply to an Object}
\usage{
findS3Meth(object)
}
\arguments{
\item{object}{An R object.}
}
\value{
A vector of applicable methods.
}
\description{
Adapted from Josh O'Brien's idea at
https://stackoverflow.com/questions/23840404/function-to-return-all-s3-methods-applicable-to-an-object
}
\examples{
data(ToothGrowth)
	tst1 <- findS3Meth(ToothGrowth); tst1
tst2 <- findS3Meth(ToothGrowth$supp); tst2
}
\keyword{utilities}

