% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HandyStuff-package.R
\docType{package}
\name{HandyStuff-package}
\alias{HandyStuff}
\alias{HandyStuff-package}
\title{Handy Functions & Stuff I Use, & You Might Too!}
\description{
Mostly unrelated functions for all sorts of things I find Handy.
}
\author{
Bryan A. Hanson, DePauw University, Greencastle Indiana USA

Maintainer: Bryan A. Hanson \email{hanson@depauw.edu}
}
\keyword{package}

