% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/levelplotPanels.R
\name{lpi}
\alias{lpi}
\alias{lpi,}
\alias{lpr}
\title{Lattice Panel Functions for Text-Decorated Levelplots}
\usage{
lpi(x, y, z, ...)
}
\arguments{
\item{x}{Numeric or integer vector, as in levelplot.}

\item{y}{Numeric or integer vector, as in levelplot.}

\item{z}{Numeric or integer vector, as in levelplot.}

\item{...}{Additional parameters to be passed to levelplot.}
}
\value{
None.  A level plot is displayed.
}
\description{
These panel functions work in conjunction with levelplot
and label the cells of the plot with text.  \code{lpi}
labels with integers and \code{lpr} labels with real
numeric (real) values.
}
\examples{
require('lattice')
# numeric test data
M1 <- abs(matrix(runif(16, 0.2, 1), nrow = 4))
rownames(M1) <- as.character(1:4)
colnames(M1) <- as.character(1:4)
# Negate the lower triangle values
M1[lower.tri(M1)] <- -1.0*M1[lower.tri(M1)]
diag(M1) <- 0.0

M2 <- t(M1[nrow(M1):1,])
levelplot(M2,
xlab.top = "column", ylab = "row", xlab = "",
scale = list(x = list(alternating = 2)),
col.regions = cm.colors(3),
at = c(-1.0, -0.01, 0.01, 1.0), panel = lpr)

# integer test data
M3 <- matrix(1:16, ncol = 4, byrow = TRUE)
rownames(M3) <- as.character(1:4)
colnames(M3) <- as.character(1:4)

M4 <- t(M3[nrow(M3):1,])
levelplot(M4,
xlab.top = "column", ylab = "row", xlab = "",
scale = list(x = list(alternating = 2)),
col.regions = heat.colors(26), panel = lpi)
}
\keyword{plot}

