% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyAOVtable.R
\name{prettyAOVtable}
\alias{prettyAOVtable}
\title{Create a nicely formated analysis of variance table as a grob}
\usage{
prettyAOVtable(aov.object = NULL, cex = 1)
}
\arguments{
\item{aov.object}{An object of class \code{aov}.}

\item{cex}{Numeric.  The relative size of text in the table.  The size of the cells
and table automatically adjust for the size of the text.}
}
\value{
A \code{gtable} / \code{grob} object which can be plotted.  See the examples.
}
\description{
Creates a nicely formated tableGrob using \code{gridExtra} and \code{broom}.
}
\examples{

require("broom")
require("grid")
require("gridExtra")
require("gtable")

npk1.aov <- aov(yield ~ block + N*P*K, npk) # from ?aov
grid.newpage()
grid.draw(prettyAOVtable(npk1.aov))
npk2.aov <- aov(yield ~ block*K, npk)
grid.newpage()
grid.draw(prettyAOVtable(npk2.aov))

}
\references{
\url{https://github.com/bryanhanson/HandyStuff}
}
\author{
Bryan A. Hanson, DePauw University. \email{hanson@depauw.edu}
}
\keyword{utilities}
