% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/screenTheme.R
\name{screenTheme}
\alias{posterTheme}
\alias{screenTheme}
\alias{screenTheme,}
\title{Lattice themes resembling ggplot2 themes}
\usage{
screenTheme()
}
\value{
A list containing the plotting theme details.
}
\description{
This is a set of theme details for use with \code{\link{compareCats}},
\code{\link{rxnNorm}} and \code{\link{lmEqn}}.  The themes are derived from package
\code{latticeExtra} with a few additional customizations.  The settings for
posterTheme are generally larger for decent viewing on a poster.
}
\author{
Bryan A. Hanson, DePauw University. \email{hanson@depauw.edu}
}
\references{
\url{http://github.com/bryanhanson/HandyStuff}
}
\seealso{
\code{\link{compareCats}} and \code{\link{rxnNorm}} for examples
of this theme in use.
}
\keyword{utilities}

