\name{screenThemeL}
\alias{screenThemeL}
\alias{posterThemeL}
\title{
Lattice themes resembling ggplot2 themes
}
\description{
This is a set of theme details for use with \code{\link{compareCatsL}} and \code{\link{rxnNormL}}.  The themes are derived from \code{\link{latticeExtra}} with a few additional customizations.  The settings for posterThemeL are generally larger for decent viewing on a poster.
}
\usage{
screenThemeL()
posterThemeL()
}
\value{
A list containing the plotting theme details.
}
\references{
\url{http://github.com/bryanhanson/HandyStuff}
}

\author{
Bryan A. Hanson, DePauw University. \email{hanson@depauw.edu}
}
\seealso{
\code{\link{compareCatsL}} and \code{\link{rxnNormL}} for examples of this theme in use.
}
\keyword{ utilities }
