% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackByColumn.R
\name{stackByColumn}
\alias{stackByColumn}
\title{Convert a Vector into a Matrix Filling Down Columns}
\usage{
stackByColumn(IN, nrow, ncol)
}
\arguments{
\item{IN}{An integer or numeric vector.}

\item{nrow}{The number of rows in the final matrix.}

\item{ncol}{The number of columns in the final matrix.}
}
\value{
A ncol x nrow matrix with no attributes
}
\description{
This is one of several related convenience functions designed to convert
vectors into matrices or matrices into vectors, either by row or by column.
}
\details{
This function is almost the same as
\code{\link{matrix}}, with \code{byrow = FALSE}, and
there are no dimnames or attributes in this function.
}
\examples{
stackByColumn(IN = 1:20, nrow = 5, ncol = 4)
}
\seealso{
\code{\link{stackByRow}}, \code{\link{vectorizeByColumn}},
\code{\link{vectorizeByRow}}
}
\keyword{utilities}
