% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sumStatDeluxe.R
\name{sumStatDeluxe}
\alias{sumStatDeluxe}
\title{Calculate Custom Summary Statistics}
\usage{
sumStatDeluxe(data = NULL, groupVars = NULL, respVars = NULL, ci = 0.95,
  ...)
}
\arguments{
\item{data}{A data frame containing the variables.}

\item{groupVars}{Character.  A vector of the categorical variable names.}

\item{respVars}{Character.  A vector of the response variable names.}

\item{ci}{Numeric.  The confidence interval desired.}

\item{...}{Other parameters to be passed downstream.}
}
\value{
Either a data frame or a list of data frames.
}
\description{
This function computes a selection of summary statistics on a data frame which
contains response variables and categorical variables.  If there is only one
response variable, the results are returned in a data frame.  For more than one
response variable, the results are returned in a list of data frames.  See the examples.
}
\examples{
require('plyr')
# One response variable returns a data frame:
tst <- sumStatDeluxe(chickwts, groupVars = "feed", respVars = "weight")
# Two response variables returns a list of data frames:
tst <- sumStatDeluxe(iris, groupVars = "Species",
respVars = c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width"))

# Two categorical variables, one response:
tst <- sumStatDeluxe(warpbreaks, groupVars = c("wool", "tension"), respVars = "breaks")
# One categorical, multiple response variables:
tst <- sumStatDeluxe(airquality, groupVars = "Month",
respVars = c("Ozone", "Solar.R", "Wind"))
}
\keyword{summary}

