% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorizeByColumn.R
\name{vectorizeByColumn}
\alias{vectorizeByColumn}
\title{Convert a Matrix into a Vector Column by Column}
\usage{
vectorizeByColumn(IN)
}
\arguments{
\item{IN}{A matrix.}
}
\value{
A vector.
}
\description{
This is one of several related convenience functions designed to convert
vectors into matrices or matrices into vectors, either by row or by column.
}
\examples{
vectorizeByColumn(matrix(1:20, ncol = 4))
}
\seealso{
\code{\link{stackByColumn}}, \code{\link{stackByRow}},
\code{\link{vectorizeByRow}}
}
\keyword{utilities}
