% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compareCats.R
\name{compareCats}
\alias{compareCats}
\title{Compare Data by Category}
\usage{
compareCats(formula = NULL, data = NULL, cols = NULL, freckles = FALSE,
  method = c("sem", "sem95", "iqr", "mad", "box", "points"),
  theme = screenTheme(), ...)
}
\arguments{
\item{formula}{A formula with both LHS and RHS.  The formula should comply
with the usual \code{lattice} conventions, for instance \code{y ~ factor1 |
factor2}.}

\item{data}{A data frame containing the response and categorical/factor
data.  A maximum two categories with any number of levels can be plotted at
one time.}

\item{cols}{A vector of colors to be used for factor 1; must be as along as
the levels in factor 1.}

\item{freckles}{Logical; if \code{TRUE}, the actual data points are plotted
as small points, hence the name, jittered slightly in the horizontal
direction only.}

\item{method}{One of \code{c("sem", "sem95", "iqr", "mad", "box",
"points")}.  Various methods for computing measures of spread and central
tendency.  See the documentation for \code{\link[ChemoSpec]{seXy}}.}

\item{theme}{Character; A suitble \code{lattice} theme.
There are two built-in themes which you can use "as is" or modify to your heart's
content.  If  none is given, \code{\link{screenTheme}} will be used.  The other option
provided is \code{\link{posterTheme}}.}

\item{\dots}{Other parameters to be passed downstream.}
}
\value{
A \code{lattice} object.  These can be modified by the usual methods
associated with these packages.
}
\description{
This function allows one to generate a nice plot comparing a response from
samples falling into one or two categories with corresponding levels.
\code{lattice} is used to create plots, with options to include the original
data and to use various measures of central tendency and spread.  A count of
the number of samples in each level is included.  Gives a visual impression
of the data to go along with hypothesis tests.
}
\details{
The supplied data frame is stripped down to just the response and factor
columns, then \code{NAs} are removed; the counts reflect this.
}
\examples{
#
### Set up test data
#
library("ChemoSpec")
mydf <- data.frame(
	resp = rnorm(40),
	cat1 = sample(LETTERS[1:3], 40, replace = TRUE),
	cat2 = sample(letters[1:2], 40, replace = TRUE))
#
### One factor:
#
p <- compareCats(formula = resp~cat1, data = mydf,
	method = "sem", freckles = TRUE, poster = FALSE,
	cols = c("red", "orange", "blue"))
print(p)
#
### Two factors:
#
p <- compareCats(formula = resp~cat1 | cat2, data = mydf,
	method = "sem", freckles = TRUE,
	cols = c("red", "orange", "blue"))
print(p)
#
### Interchange the roles of the factors
#
p <- compareCats(formula = resp~cat2 | cat1, data = mydf,
	method = "sem", freckles = TRUE,
	cols = c("red", "blue"))
print(p)
}
\author{
Bryan A. Hanson, DePauw University. \email{hanson@depauw.edu}
}
\references{
\url{http://github.com/bryanhanson/HandyStuff}
}
\keyword{plot}
\keyword{univariate}

