% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggqq_norm.R
\name{ggqq_norm}
\alias{ggqq_norm}
\title{Draw a QQ Plot using ggplot2}
\usage{
ggqq_norm(dataframe = NULL, respvar = NULL, facetvars = NULL)
}
\arguments{
\item{dataframe}{A data frame containing the variables to be plotted.  \code{NA}  will
be omitted by this function.}

\item{respvar}{Character.  The name of the response variable in the data frame.}

\item{facetvars}{Character.  One or two names of the variables to be used for facetting.}
}
\value{
A ggplot2 object.  This can be printed by \code{print(object)} or just \code{object}.
}
\description{
This function creates a QQ plot (normal probability plot), including a reference line,
using ggplot2 graphics.  The reference line is based on the normal distribution.
The plot may be faceted.
}
\examples{
ns <- 200
require("ggplot2")
#
td1 <- data.frame(resp = rnorm(ns), f1 = sample(c("A", "B"), ns, replace = TRUE),
f2 = sample(c("C", "D"), ns, replace = TRUE))
res <- ggqq_norm(td1, "resp", c("f1", "f2"))
print(res + labs(title = "Normal Distribution"))
#
td1 <- data.frame(resp = rnorm(ns), f1 = sample(c("A", "B"), ns, replace = TRUE),
f2 = sample(c("C", "D"), ns, replace = TRUE))
bad <- sample(1:ns, 4)
td1$resp[bad[1:2]] <- 5
td1$resp[bad[3:4]] <- -5
res <- ggqq_norm(td1, "resp", c("f1", "f2"))
print(res + labs(title = "Normal Distribution + Outliers"))
#
td2 <- data.frame(resp = runif(ns), f1 = sample(c("A", "B"), ns, replace = TRUE),
f2 = sample(c("C", "D"), ns, replace = TRUE))
res <- ggqq_norm(td2, "resp", c("f1", "f2"))
print(res + labs(title = "Uniform Distribution"))
}
\keyword{plot}

