\name{HandyStuff-package}
\alias{HandyStuff-package}
\alias{HandyStuff}
\docType{package}
\title{
Handy Functions & Stuff I Use, & You Might Too!
}
\description{
Mostly unrelated functions for all sorts of things I want to do or I want my students to be able to do easily.  Also a few data sets I like to use in teaching.  And other fun stuff.  An incoherent collection of seemingly random functions.
}
\details{
\tabular{ll}{
Package: \tab HandyStuff\cr
Type: \tab Package\cr
Version: \tab 0.3\cr
Date: \tab 2011-05-23\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
}
\author{
Bryan A. Hanson, DePauw University, Greencastle Indiana USA

Maintainer: Bryan A. Hanson \email{hanson@depauw.edu}
}
\keyword{ package }
