% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lmEqn.R
\name{lmEqn}
\alias{lmEqn}
\title{Plot Points, Fit a Line, Display Equation}
\usage{
lmEqn(formula = NULL, data = NULL, method = "lm", leg.loc = c(0.5, 0.5),
  theme = screenTheme(), ...)
}
\arguments{
\item{formula}{A formula with both LHS and RHS.  The formula should comply
with the usual \code{R} standards, for instance \code{y ~ x}.}

\item{data}{A data frame containing two columns whose names correspond to
the specified formula.}

\item{method}{A character string.  Either "lm" for linear model or "rlm" for
a robust fit.}

\item{leg.loc}{A vector of two numbers (x, y) giving the location of the
legend.  The values correspond to the units of the data. See the examples.}

\item{theme}{Character; A suitble \code{lattice} theme.
There are two built-in themes which you can use "as is" or modify to your heart's
content.  If  none is given, \code{\link{screenTheme}} will be used.  The other option
provided is \code{\link{posterTheme}}.}

\item{\dots}{Additional arguments to be passed downstream.}
}
\value{
A plot is drawn and the \code{ggplot2} or \code{lattice} object is
returned, possibly for further manipulation.
}
\description{
This function plots the points given, then fits a line and displays the
regression statistics using \code{lattice}.
}
\examples{
conc = seq(1, 12, length.out = 8)
abs = jitter(conc, factor = 2)*0.1
cc <- data.frame(conc, abs)
lmEqn(formula = abs ~ conc, data = cc, method = "rlm",
leg.loc = c(3, 0.1),
xlab = "concentration", ylab = "absorbance",
main = "Calibration Curve")
}
\author{
Bryan A. Hanson, DePauw University. \email{hanson@depauw.edu}
}
\keyword{plot}
\keyword{utilities}

