% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rxnNorm.R
\name{rxnNorm}
\alias{rxnNorm}
\title{Create a Reaction Norm or Interaction Plot}
\usage{
rxnNorm(formula = NULL, data = NULL, groups = NULL, cols = NULL,
  freckles = FALSE, type = "connect", method = c("sem", "sem95", "iqr",
  "mad"), table = NULL, theme = screenTheme(), ...)
}
\arguments{
\item{formula}{A formula with both LHS and RHS.  The formula should comply
with the usual \code{lattice} conventions, for instance \code{y ~ x_factor}.}

\item{data}{A data frame containing the response and categorical/factor
data, and possibly group data.}

\item{groups}{The (unquoted) name of a variable in \code{data} to serve as
the grouping variable.}

\item{cols}{A vector of colors to be used for factor 1; must be as along as
the levels in factor 1.}

\item{freckles}{Logical; if \code{TRUE}, the actual data points are plotted
as small points, hence the name, jittered slightly in the horizontal
direction only.}

\item{type}{Either "connect", "fitLine" or "anova".  In the first case, a
line is drawn which connects the chosen measure of central tendency (e.g.
mean) for each value of \code{fac1}.  This would generally be used when the
x value are categorical.  In the second case, a linear fit is performed and
the line is drawn; used for continuous x values.  In the third case, a
connecting line is drawn along with an ANOVA summary table.}

\item{method}{One of \code{c("sem", "sem95", "iqr", "mad")}.  Each produces
a dot at the measure of central tendency and error bars corresponding to the
requested method ("box" currently does not work quite right).  Full details
are in \code{\link[ChemoSpec]{seXy}}.}

\item{table}{If NULL, the default, no action.  If a vector of 3 numbers
\code{c(x, y, cex)}, a summary table is placed at the given coordinates with
font size cex.  The table contains either a summary of counts (if \code{type
= "connect"}), a report of the linear fit (if \code{type = "fitLine"}) or an
ANOVA table (if \code{type = "anova"}.  The coordinates are in device units
so 0.5, 0.5 puts the table dead center in the plotting region.}

\item{theme}{Character; A suitable \code{lattice} theme.
There are two built-in themes which you can use "as is" or modify to your heart's
content.  If  none is given, \code{\link{screenTheme}} will be used.  The other option
provided is \code{\link{posterTheme}}.}

\item{\dots}{Additional arguments to be passed downstream.}
}
\value{
A \code{lattice} plot object is returned invisibly, and a plot is
created.
}
\description{
Draws a reaction norm or interaction plot as typically used in ecology to
demonstrate treatment x genotype interactions.  The x axis can be
categorical or numeric.  There are options to plot the raw data points in
addition to a choice of several summary methods.  The central values may be
connected or the entire data set fitted to a line.  A summary table is
optional, and gives either counts at each level, an ANOVA table or fit data.
}
\details{
The choices in \code{method} are based upon functions in the
\code{\link[ChemoSpec]{ChemoSpec}} package which will need to be installed.
See \code{\link[ChemoSpec]{seXy}} for the details.
}
\examples{
#
### set up demo data
#
require("ChemoSpec")
require("lattice")
require("gridExtra")
set.seed(79)
res = c(rnorm(10, 5, 1.5), rnorm(10, 8, 2),
rnorm(10, 14, 2.0), rnorm(10, 10, 1.5),
rnorm(10, 15, 2), rnorm(10, 12,2.5))
fac1 <- c(rep("L", 20), rep("M", 20), rep("H", 20))
fac1 <- factor(fac1, levels = c("L", "M", "H"))
fac2 <- sample(c("A", "B"), 60, replace = TRUE)
fac2 <- as.factor(fac2)
num <- c(rep(5, 20), rep(10, 20), rep(12, 20))
td <- data.frame(resp = res, fac1 = fac1, fac2 = fac2, num = num)
#
### Numeric x with lm and summary table:
#
p <- rxnNorm(formula = resp~num, groups = fac2, data = td,
method = "iqr", freckles = TRUE, type = "fitLine",
cols = c("red", "blue"), table = c(0.5, 0.3, 0.75),
main = "rxnNorm - linear model w/fit table")
print(p)
#
### Categorical x, connected and summary table:
#
p <- rxnNorm(formula = resp~fac1, groups = fac2, data = td,
method = "iqr", freckles = TRUE, type = "connect",
cols = c("red", "blue"), table = c(0.5, 0.3, 0.75),
main = "rxnNorm - Categorical x w/summary table")
print(p)
#
### Categorical x, connected with ANOVA table:
#
p <- rxnNorm(formula = resp~fac1, groups = fac2, data = td,
method = "iqr", freckles = TRUE, type = "anova",
cols = c("red", "blue"), table = c(0.6, 0.2, 0.75),
main = "rxnNorm - Categorical x w/ANOVA table")
print(p)
}
\author{
Bryan A. Hanson, DePauw University. \email{hanson@depauw.edu}
}
\references{
\url{http://github.com/bryanhanson/HandyStuff}
}
\keyword{plot}
\keyword{univariate}

